fix(client): use correct user-agent in actor fetches
  The ActorFetcher was using a hardcoded User-Agent ("apkit/0.3.0")
  instead of the one configured on the client instance.

  This change centralizes header processing to a shared utility that
  correctly applies the client's user_agent. The actor methods now use
  this utility, ensuring the correct User-Agent is sent.

  This also adds the ability to pass custom headers to these methods.

  Fixes: #34
