+++ src/apkit/client/asyncio/actor.py
@@ -1,6 +1,7 @@
 from typing import TYPE_CHECKING, Union
 
 from apmodel.types import ActivityPubModel
+
 from .. import _common, models
 
 if TYPE_CHECKING:
@@ -11,12 +12,20 @@ class ActorFetcher:
     def __init__(self, client: "ActivityPubClient"):
         self.__client: "ActivityPubClient" = client
 
-    async def resolve(self, username: str, host: str) -> models.WebfingerResult:
+    async def resolve(
+        self,
+        username: str,
+        host: str,
+        headers: dict = {"Accept": "application/jrd+json"},
+    ) -> models.WebfingerResult:
         """Resolves an actor's profile from a remote server asynchronously."""
+        headers = _common.ensure_user_agent_and_reconstruct(
+            headers, self.__client.user_agent
+        )
         resource = models.Resource(username=username, host=host)
         url = _common.build_webfinger_url(host=host, resource=resource)
 
-        async with self.__client.get(url, headers={"User-Agent": "apkit/0.3.0"}) as resp:
+        async with self.__client.get(url, headers=headers) as resp:
             if resp.ok:
                 data = await resp.json()
                 result = models.WebfingerResult.from_dict(data)
@@ -25,8 +34,13 @@ class ActorFetcher:
             else:
                 raise ValueError(f"Failed to resolve Actor: {url}")
 
-    async def fetch(self, url: str) -> Union[ActivityPubModel, dict]:
-        async with self.__client.get(url, headers={"User-Agent": "apkit/0.3.0", "Accept": "application/activity+json"}) as resp:
+    async def fetch(
+        self, url: str, headers: dict = {"Accept": "application/activity+json"}
+    ) -> Union[ActivityPubModel, dict]:
+        headers = _common.ensure_user_agent_and_reconstruct(
+            headers if headers else {}, self.__client.user_agent
+        )
+        async with self.__client.get(url, headers=headers) as resp:
             if resp.ok:
                 data = await resp.parse()
                 return data
