+++ src/apkit/client/asyncio/client.py
@@ -6,7 +6,6 @@ from typing import (
     Any,
     Awaitable,
     Callable,
-    Dict,
     Iterable,
     List,
     Literal,
@@ -40,56 +39,9 @@ from yarl import URL, Query
 
 from ..._version import __version__
 from ...types import ActorKey
-from .._common import sign_request
+from .._common import ensure_user_agent_and_reconstruct, sign_request
 from .actor import ActorFetcher
 from .types import ActivityPubClientResponse, _RequestContextManager
-
-
-def ensure_user_agent_and_reconstruct(
-    headers: LooseHeaders, user_agent: str
-) -> Dict[str, str]:
-    processed_headers: Dict[str, str] = {}
-
-    if isinstance(headers, Mapping):
-        for k, v in headers.items():
-            key_str = str(k)
-            key_lower = key_str.lower()
-
-            if key_lower not in processed_headers:
-                processed_headers[key_lower] = v
-                processed_headers[key_lower + "_original_key"] = key_str
-
-    elif isinstance(headers, Iterable):
-        for k, v in headers:
-            key_str = str(k)
-            key_lower = key_str.lower()
-
-            if key_lower not in processed_headers:
-                processed_headers[key_lower] = v
-                processed_headers[key_lower + "_original_key"] = key_str
-
-    else:
-        raise TypeError(f"Unsupported header type: {type(headers)}")
-
-    user_agent_key_lower = "user-agent"
-    if user_agent_key_lower not in processed_headers:
-        processed_headers[user_agent_key_lower] = user_agent
-        processed_headers[user_agent_key_lower + "_original_key"] = "User-Agent"
-
-    final_headers: Dict[str, str] = {}
-    for key_lower, value in processed_headers.items():
-        if key_lower.endswith("_original_key"):
-            continue
-
-        original_key = processed_headers.get(key_lower + "_original_key")
-
-        if original_key:
-            final_headers[original_key] = value
-        else:
-            standard_key = key_lower.replace("-", " ").title().replace(" ", "-")
-            final_headers[standard_key] = value
-
-    return final_headers
 
 
 class ActivityPubClient(aiohttp.ClientSession):
@@ -325,7 +277,11 @@ class ActivityPubClient(aiohttp.ClientSession):
         signatures: List[ActorKey] = [],
         sign_with: Optional[
             List[Literal["draft-cavage", "rsa2017", "fep8b32", "rfc9421"]]
-        ] = ["draft-cavage", "rsa2017", "fep8b32"],  # TODO: "draft-cavage", "rsa2017", "fep8b32"
+        ] = [
+            "draft-cavage",
+            "rsa2017",
+            "fep8b32",
+        ],  # TODO: "draft-cavage", "rsa2017", "fep8b32"
         # deprecated
         key_id: Optional[str] = None,
         signature: Optional[
